+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add message on invalid routing response (#764)
 - Consistent exception type on invalid profile (#771)
 - Pass zero amount directly instead of its size (#776)
+- Add named constants for default threads number and exploration level (#805)
 
 ### Fixed